+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [1.7.0] - 2026-05-14
+
 ### Added
 
 - Add option `--portable` to guard generated MSVC-specific CMake settings so they are inactive for other compilers.
@@ -122,7 +124,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 First stable release.
 
-[unreleased]: https://github.com/chausner/vcxproj2cmake/compare/v1.6.0...HEAD
+[unreleased]: https://github.com/chausner/vcxproj2cmake/compare/v1.7.0...HEAD
+[1.7.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.3.0...v1.4.0